VASP: Encapsulate the output parser in the parse_vasp_output() function
